refactor: update package structure (invent API pkg)
