+++ internal/opt/api/rest/manager_test.go
@@ -1,4 +1,4 @@
-package moderunner
+package rest
 
 import (
 	"log/slog"
