+++ internal/opt/api/rest/receivemode/model.go
