+++ internal/opt/api/supervisors/archivesv/upload.go
@@ -1,4 +1,4 @@
-package receivesuperv
+package archivesv
 
 import (
 	"context"
