+++ internal/opt/api/supervisors/archivesv/upload_test.go
@@ -1,4 +1,4 @@
-package receivesuperv
+package archivesv
 
 import (
 	"context"
