+++ internal/opt/api/supervisors/backupsv/retain.go
@@ -1,4 +1,4 @@
-package backupsuperv
+package backupsv
 
 import (
 	"slices"
