+++ internal/opt/api/supervisors/backupsv/retain_test.go
@@ -1,4 +1,4 @@
-package backupsuperv
+package backupsv
 
 import (
 	"testing"
