+++ internal/opt/basebackup/restore_info.go
@@ -4,10 +4,11 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	backupdto2 "github.com/pgrwl/pgrwl/internal/opt/basebackup/backupdto"
 	"log/slog"
 	"path/filepath"
 	"strings"
+
+	backupdto2 "github.com/pgrwl/pgrwl/internal/opt/basebackup/backupdto"
 
 	st "github.com/pgrwl/pgrwl/internal/opt/shared/storecrypt"
 )
