+++ internal/opt/basebackup/stream_bb.go
@@ -6,11 +6,12 @@ import (
 	"encoding/binary"
 	"encoding/json"
 	"fmt"
-	"github.com/pgrwl/pgrwl/internal/opt/basebackup/backupdto"
 	"io"
 	"log/slog"
 	"strings"
 	"time"
+
+	"github.com/pgrwl/pgrwl/internal/opt/basebackup/backupdto"
 
 	"github.com/pgrwl/pgrwl/internal/opt/metrics/backupmetrics"
 	"github.com/pgrwl/pgrwl/internal/opt/shared/x/fsx"
